+++ lib/glific/templates.ex
@@ -14,7 +14,6 @@ defmodule Glific.Templates do
     Providers.Gupshup.Template,
     Repo,
     Settings,
-    Settings.Language,
     Tags.Tag,
     Tags.TemplateTag,
     Templates.SessionTemplate
@@ -570,55 +569,4 @@ defmodule Glific.Templates do
     do: button["text"] <> ", " <> button["phone_number"]
 
   defp do_parse_buttons("QUICK_REPLY", button), do: button["text"]
-
-  @doc """
-  Import pre approved templates when BSP is GupshupEnterprise
-  """
-  @spec import_enterprise_templates(non_neg_integer(), String.t()) :: {:ok, any}
-  def import_enterprise_templates(organization_id, data) do
-    {:ok, stream} = StringIO.open(data)
-    organization = Partners.organization(organization_id)
-
-    stream
-    |> IO.binstream(:line)
-    |> CSV.decode(headers: true, strip_fields: true)
-    |> Enum.map(fn {_, data} -> import_approved_templates(data) end)
-    |> do_update_hsms(organization, :gupshup_enterprise)
-
-    {:ok, %{message: "All templates have been added"}}
-  end
-
-  @spec import_approved_templates(map()) :: map()
-  defp import_approved_templates(template),
-    do: %{
-      "id" => Ecto.UUID.generate(),
-      "data" => template["Body"],
-      "meta" => get_example_body(template["Body"]),
-      "category" => "ALERT_UPDATE",
-      "elementName" => template["Template Name"],
-      "languageCode" => get_language(template["Language"]),
-      "templateType" => template["Type"],
-      "status" => get_status(template["Status"]),
-      "enterprise_id" => template["Template Id"]
-    }
-
-  @spec get_language(String.t()) :: String.t()
-  defp get_language(label_locale) do
-    {:ok, language} = Repo.fetch_by(Language, %{label_locale: label_locale})
-    language.locale
-  end
-
-  @spec get_status(String.t()) :: String.t()
-  defp get_status("Enabled"), do: "APPROVED"
-  defp get_status("Rejected"), do: "REJECTED"
-  defp get_status(_status), do: "PENDING"
-
-  @spec get_example_body(String.t()) :: String.t()
-  defp get_example_body(body) do
-    body
-    |> String.replace("{{", "[sample text ")
-    |> String.replace("}}", "]")
-    |> then(&Map.put(%{}, :example, &1))
-    |> Jason.encode!()
-  end
 end
