+++ test/glific/templates_test.exs
@@ -3,6 +3,7 @@ defmodule Glific.TemplatesTest do
 
   alias Glific.{
     Fixtures,
+    Providers.GupshupEnterprise.Template,
     Seeds.SeedsDev,
     Settings,
     Templates,
@@ -998,7 +999,7 @@ defmodule Glific.TemplatesTest do
       data =
         "Template Id,Template Name,Body,Type,Quality Rating,Language,Status,Created On\r\n6122571,2meq_payment_link,	Your OTP for {{1}} is {{2}}. This is valid for {{3}}.,TEXT,Unknown,English,Enabled,2021-07-16\n6122572,meq_payment_link2,You are one step away! Please click the link below to make your payment for the Future Perfect program.,TEXT,Unknown,English,Rejected,2021-07-16"
 
-      Templates.import_enterprise_templates(attrs.organization_id, data)
+      Template.import_enterprise_templates(attrs.organization_id, data)
 
       assert {:ok, %SessionTemplate{} = imported_template} =
                Repo.fetch_by(SessionTemplate, %{enterprise_template_id: "6122571"})
